% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viz_MA_data}
\alias{viz_MA_data}
\title{Meta-analytic data to visualize using viz_MARC()}
\format{
An object of class \code{"data.frame"} with 180 rows and 5 columns
\describe{
 \item{ccode}{ISO3 country code (as character) for the countries in the
     sample (Angola, Benin, France, Rwanda, and UK)}
 \item{year}{A value between 1990 and 1999}
 \item{month}{An abbreviation (MMM) for month (character)}
 \item{gpd}{A fake value for GDP (randomly generated)}
 \item{population}{A fake value for population (randomly generated)}
}
}
\usage{
data(viz_MA_data)
}
\description{
Includes data for 4 meta-analyses of size k = 10, 20, 50, 100.
Should filter/subset by the first column (k) to extract data for one meta-analysis
}
\examples{

data(viz_MA_data)
head(viz_MA_data)

}
\references{
This data set was generated for the statistical cognition experiment in Fitzgerald, Khella, Charles, & Tipton (2024).
}
\keyword{datasets}
