% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_MARC.R
\name{viz_MARC}
\alias{viz_MARC}
\title{viz_MARC}
\usage{
viz_MARC(
  d_j = NULL,
  se_j = NULL,
  study_labels = NULL,
  seed = NULL,
  x_limits = NULL,
  confidence_level = 0.95
)
}
\arguments{
\item{d_j}{vector of effect size estimates}

\item{se_j}{vector of standard errors of effect size estimates}

\item{study_labels}{vector of study labels (optional)}

\item{seed}{integer used to set random seed before randomizing rows (optional)}

\item{x_limits}{vector of x-axis limits c(xmin, xmax) to be used in plotting (optional)}

\item{confidence_level}{confidence level for interval written in plot annotation (default = 0.95)}
}
\value{
a plotly object
}
\description{
Produces an interactive Meta-Analytic Rain Cloud (MARC) plot
}
\examples{
data(viz_MA_data)
d_j <- viz_MA_data \%>\% filter(k == 100) \%>\% pull(d_j)
se_j <- viz_MA_data \%>\% filter(k == 100) \%>\% pull(se_j)
viz_MARC(d_j, se_j, seed = 437)
}
