% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_MARC.R
\name{viz_MARC}
\alias{viz_MARC}
\title{viz_MARC}
\arguments{
\item{d_j}{vector of effect size estimates}

\item{se_j}{vector of standard errors of effect size estimates}

\item{w_j}{vector or matrix of user input weights for the effect size estimates}

\item{method_obj}{metafor object used to fit the meta-analytic model, includes rma.uni(),rma(), rma.mh(),rma.peto(), rma.glmm() or rma.mv() (If not specified, rma.uni(yi=d_j,sei=se_j, method="FE") is used)}

\item{summary_es}{user input meta-analytic summary effect size}

\item{summary_se}{user input meta-analytic summary standard error}

\item{confidence_level}{confidence level for interval written in plot annotation (default = 0.95)}

\item{summary_only}{TRUE/FALSE indicator for whether to display the summary effect ONLY (default = FALSE)}

\item{study_labels}{vector of study labels (optional)}

\item{show_study_labels}{TRUE/FALSE indicator for whether to display study labels (default = FALSE)}

\item{seed}{integer used to set random seed before randomizing rows (optional)}

\item{x_limits}{vector of x-axis limits c(xmin, xmax) to be used in plotting (optional)}

\item{y_limits}{vector of y-axis limits c(ymin, ymax) to be used in plotting (optional)}

\item{y_limits_rect}{vector of y-axis limits c(ymin, ymax) to be used in plotting white rectangle (optional)}

\item{max_dot_size}{maximum dot size, used in ggplot static version (default = 10, reduce if summary too big)}

\item{width_in}{desired width (in inches) for scaling purposes only}

\item{height_in}{desired height (in inches) for scaling purposes only}

\item{textbox_width}{controls width of textbox (default = 4 inches)}

\item{font_sizes}{vector of 9 font sizes for annotations, in order: SUMMARY OF EVIDENCE, Average SMD & Weight, CI annotation, Increased/Decreased Scores, More/Less certain, Legend number labels, Legend Weight title, X-axis label, Y-axis label (default = c(14, 10, 8, 9, 8, 7, 10,16,16))}

\item{digits}{number of digits to round summary effect size and confidence interval bounds to (default = 2)}

\item{xinc}{controls the x axis increments for the standardized mean differences (default = .2)}

\item{dot_color}{controls the color of the effect size dots using ggplot colors (default=navyblue)}

\item{dot_trans}{controls the transparency of the effect size dots on a scacle of 0 to 1 (default = .5)}

\item{font_type}{controls the font type for the text (default = "")}
}
\value{
a ggplot object
}
\description{
Produces an static Meta-Analytic Rain Cloud (MARC) plot
}
\examples{
library(tidyverse)
data(viz_MA_data)
d_j <- viz_MA_data \%>\% filter(k == 100) \%>\% pull(d_j)
se_j <- viz_MA_data \%>\% filter(k == 100) \%>\% pull(se_j)
viz_MARC(d_j, se_j, seed = 437)
viz_MARC(d_j, se_j, seed = 437, type = "static")
}
